transportation/requirements.txt
requests
Pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
tabula-py
pandas